[Upd] moved finder views from www to catalog
